refactor<app/launch>: 去掉启动节点时默认添加的 use_sim_time 参数
