+++ laser_copilot_applications/launch/visualization.launch.py
@@ -1,0 +1,27 @@
+import launch
+from launch_ros.actions import ComposableNodeContainer
+from launch_ros.descriptions import ComposableNode
+
+
+def get_composable_node():
+    return ComposableNode(
+        package="laser_copilot_applications",
+        plugin="laser_copilot_applications::visualization_helper",
+        name="visualization_helper",
+        parameters=[{}],
+        remappings=[],
+    )
+
+
+def generate_launch_description():
+    """Generate launch description with multiple components."""
+    container = ComposableNodeContainer(
+        name="laser_copilot_container",
+        namespace="",
+        package="rclcpp_components",
+        executable="component_container",
+        composable_node_descriptions=[get_composable_node()],
+        output="screen",
+    )
+
+    return launch.LaunchDescription([container])
